requirement.txt
algorithmia==1.0.7
certifi==2017.4.17
enum34==1.1.6
future==0.16.0
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
urllib3==1.20
